refactor: 회원 엔티티 이름 변경, 패키지 정리 #39
