--- server/src/main/java/com/hobbyt/chat/entity/ChatMessage.java
@@ -1,34 +1,0 @@
-package com.hobbyt.chat.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-
-import com.hobbyt.global.entity.BaseEntity;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class ChatMessage extends BaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-	@Column
-	private String content;
-	@Column
-	private String image;
-	@Column(nullable = false)
-	private boolean modified = false;
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "chat_user_id")
-	private ChatUser chatUser;
-}
