+++ server/src/main/java/com/hobbyt/domain/member/entity/Authority.java
@@ -1,0 +1,6 @@
+package com.hobbyt.domain.member.entity;
+
+public enum Authority {
+	ROLE_ADMIN,
+	ROLE_USER
+}
