+++ server/src/main/java/com/hobbyt/domain/member/entity/MemberStatus.java
@@ -1,0 +1,6 @@
+package com.hobbyt.domain.member.entity;
+
+public enum MemberStatus {
+	MEMBER,
+	WITHDRAWAL
+}
