+++ server/src/main/java/com/hobbyt/domain/member/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.hobbyt.domain.member.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.hobbyt.domain.member.entity.Member;
+
+public interface UserRepository extends JpaRepository<Member, Long> {
+	Optional<Member> findByEmail(String email);
+
+	boolean existsByEmail(String email);
+}
