+++ server/src/main/java/com/hobbyt/domain/member/service/NotificationEmail.java
@@ -1,0 +1,17 @@
+package com.hobbyt.domain.member.service;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class NotificationEmail {
+	private String receiver;
+	private String subject;
+	private String content;
+
+	public static NotificationEmail of(String receiver, String subject, String content) {
+		return new NotificationEmail(receiver, subject, content);
+	}
+}
