+++ server/src/main/java/com/hobbyt/domain/notification/entity/Notification.java
@@ -1,7 +1,4 @@
-package com.hobbyt.domain.entity;
-
-import java.util.ArrayList;
-import java.util.List;
+package com.hobbyt.domain.notification.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -13,32 +10,31 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
 
 import com.hobbyt.domain.member.entity.Member;
 import com.hobbyt.global.entity.BaseEntity;
 
 import lombok.AccessLevel;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "orders")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Order extends BaseEntity {
+@Getter
+public class Notification extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(nullable = false, updatable = false)
 	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "member_id", nullable = false)
-	private Member member;
-
+	@Column(nullable = false, updatable = false, name = "sender_nickname")
+	private String sender;
+	@Column(nullable = false, updatable = false)
 	@Enumerated(EnumType.STRING)
+	private NotificationType type;
+	@Column(nullable = false, updatable = false, name = "article_title")
+	private String title;
 	@Column(nullable = false)
-	private OrderStatus status;
-
-	@OneToMany(mappedBy = "order")
-	List<OrderItem> orderItems = new ArrayList<>();
+	private boolean checked = false;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "reciever_id")
+	private Member member;
 }
