+++ server/src/main/java/com/hobbyt/domain/notification/entity/NotificationType.java
@@ -1,0 +1,6 @@
+package com.hobbyt.domain.notification.entity;
+
+public enum NotificationType {
+	POST,
+	SALE
+}
