+++ server/src/main/java/com/hobbyt/domain/post/entity/Post.java
@@ -1,7 +1,5 @@
-package com.hobbyt.domain.entity;
+package com.hobbyt.domain.post.entity;
 
-import javax.persistence.Column;
-import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -14,27 +12,17 @@ import com.hobbyt.domain.member.entity.Member;
 import com.hobbyt.global.entity.Article;
 
 import lombok.AccessLevel;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Sale extends Article {
+@Getter
+public class Post extends Article {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(nullable = false, updatable = false)
 	private Long id;
-
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "writer_id")
-	private Member writer;
-
-	@Column(nullable = false, columnDefinition = "MEDIUMTEXT")
-	private String refundPolicy;
-
-	@Embedded
-	@Column(nullable = false)
-	private Period period;
-
-	@Column(nullable = false)
-	private Account account;
+	private Member member;
 }
