+++ server/src/main/java/com/hobbyt/domain/post/entity/PostComment.java
@@ -1,44 +1,34 @@
-package com.hobbyt.domain.entity;
-
-import java.util.ArrayList;
-import java.util.List;
+package com.hobbyt.domain.post.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
 
 import com.hobbyt.domain.member.entity.Member;
 import com.hobbyt.global.entity.BaseEntity;
 
 import lombok.AccessLevel;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "orders")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Order extends BaseEntity {
+@Getter
+public class PostComment extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(nullable = false, updatable = false)
 	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "member_id", nullable = false)
-	private Member member;
-
-	@Enumerated(EnumType.STRING)
 	@Column(nullable = false)
-	private OrderStatus status;
-
-	@OneToMany(mappedBy = "order")
-	List<OrderItem> orderItems = new ArrayList<>();
+	private String content;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "writer_id")
+	private Member member;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "post_id")
+	private Post post;
 }
