--- server/src/main/java/com/hobbyt/domain/user/controller/UserController.java
@@ -1,28 +1,0 @@
-package com.hobbyt.domain.user.controller;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.hobbyt.domain.user.dto.request.SignupRequest;
-import com.hobbyt.domain.user.service.UserService;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequestMapping("/api/users")
-@RequiredArgsConstructor
-public class UserController {
-
-	private final UserService userService;
-
-	@PostMapping("/signup")
-	public ResponseEntity signup(@Validated @RequestBody SignupRequest signupRequest) {
-		Long id = userService.createUser(signupRequest);
-		return new ResponseEntity(id, HttpStatus.CREATED);
-	}
-}
