--- server/src/main/java/com/hobbyt/domain/user/entity/Authority.java
@@ -1,6 +1,0 @@
-package com.hobbyt.domain.user.entity;
-
-public enum Authority {
-	ROLE_ADMIN,
-	ROLE_USER
-}
