--- server/src/main/java/com/hobbyt/domain/user/entity/UserStatus.java
@@ -1,6 +1,0 @@
-package com.hobbyt.domain.user.entity;
-
-public enum UserStatus {
-	MEMBER,
-	WITHDRAWAL
-}
