--- server/src/main/java/com/hobbyt/domain/user/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.hobbyt.domain.user.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.hobbyt.domain.user.entity.User;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-	Optional<User> findByEmail(String email);
-
-	boolean existsByEmail(String email);
-}
