--- server/src/main/java/com/hobbyt/domain/user/service/MailService.java
@@ -1,46 +1,0 @@
-package com.hobbyt.domain.user.service;
-
-import org.springframework.mail.MailException;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.mail.javamail.MimeMessagePreparator;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Service;
-
-import com.hobbyt.global.error.exception.CMailException;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class MailService {
-	private static final String ENCODING = "UTF-8";
-
-	private final JavaMailSender mailSender;
-
-	@Async
-	public void sendMail(final NotificationEmail notificationEmail) {
-		MimeMessagePreparator messagePreparator = toMimeMessagePreparator(notificationEmail);
-
-		try {
-			log.info("MailService Thread: " + Thread.currentThread().getName());
-			mailSender.send(messagePreparator);
-		} catch (MailException e) {
-			throw new CMailException();
-		}
-	}
-
-	private MimeMessagePreparator toMimeMessagePreparator(NotificationEmail notificationEmail) {
-		MimeMessagePreparator messagePreparator = mimeMessage -> {
-			MimeMessageHelper messageHelper = new MimeMessageHelper(mimeMessage, true, ENCODING);
-
-			messageHelper.setTo(notificationEmail.getReceiver());
-			messageHelper.setSubject(notificationEmail.getSubject());
-			messageHelper.setText(notificationEmail.getContent(), true);
-		};
-
-		return messagePreparator;
-	}
-}
