--- server/src/main/java/com/hobbyt/domain/user/service/NotificationEmail.java
@@ -1,17 +1,0 @@
-package com.hobbyt.domain.user.service;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class NotificationEmail {
-	private String receiver;
-	private String subject;
-	private String content;
-
-	public static NotificationEmail of(String receiver, String subject, String content) {
-		return new NotificationEmail(receiver, subject, content);
-	}
-}
