--- server/src/main/java/com/hobbyt/notification/entity/NotificationType.java
@@ -1,6 +1,0 @@
-package com.hobbyt.notification.entity;
-
-public enum NotificationType {
-	POST,
-	SALE
-}
